Fix NSRangeException crash
When a new note is created and the user presses back really fast, it will cause a NSRangeException. The root cause is that the table view has not reloaded yet and the noteChanged function is trying to scroll to an index in the tableview that hasnt existed.

The fix is to simply force a refresh when we add a note in the noteChanged method before it tries to scroll to that index.